fix: デザインパターン変更後コミット
